refactor(station-service): 서비스 레이어 추가
