+++ src/main/java/wooteco/subway/line/controller/LineController.java
@@ -5,13 +5,12 @@ import org.slf4j.LoggerFactory;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
-import wooteco.subway.line.dao.LineDao;
 import wooteco.subway.line.domain.Line;
 import wooteco.subway.line.dto.LineResponse;
 import wooteco.subway.line.dto.LineSaveRequestDto;
 import wooteco.subway.line.dto.LineUpdateRequest;
 import wooteco.subway.line.service.LineService;
-import wooteco.subway.service.ResponseError;
+import wooteco.subway.ResponseError;
 
 import java.net.URI;
 import java.util.Collections;
