+++ src/main/java/wooteco/subway/station/service/StationService.java
@@ -1,0 +1,38 @@
+package wooteco.subway.station.service;
+
+import org.springframework.stereotype.Service;
+import wooteco.subway.station.dao.StationDao;
+import wooteco.subway.station.domain.Station;
+
+import java.util.List;
+
+@Service
+public class StationService {
+    private final StationDao stationDao;
+
+    public StationService(StationDao stationDao) {
+        this.stationDao = stationDao;
+    }
+
+    public Station save(Station station) {
+        return stationDao.save(station);
+    }
+
+    public List<Station> findAll() {
+        return stationDao.findAll();
+    }
+
+    public Station findById(Long id) {
+        return stationDao.findById(id).orElseThrow(() ->
+                new IllegalArgumentException("[ERROR] 존재하지 않는 역입니다."));
+    }
+
+    public Station findByName(String name) {
+        return stationDao.findByName(name).orElseThrow(() ->
+                new IllegalArgumentException("[ERROR] 존재하지 않는 역입니다."));
+    }
+
+    public void delete(Long id) {
+        stationDao.delete(id);
+    }
+}
